requirements.txt
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
django-environ==0.12.0
boto3==1.38.46
botocore==1.38.46